+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/MainWindowVM.cs
@@ -1,4 +1,5 @@
-﻿using TD.AppInfra.Services;
+﻿using TD.AppInfra.DesignTime;
+using TD.AppInfra.Services;
 using TD.RepositoryContracts;
 using TD.WPF.Commands;
 using TD.WPF.Services;